expert-agent/pyproject.toml
[project]
name = "expert-agent"
version = "0.1.0"
description = "A2A - Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.5",
    "click>=8.1.8",
    "fastmcp>=1.0",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.4.9",
    "mcp[cli]>=1.5.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "langchain_community>=0.3.26",
    "requests==2.31.0",
    "python-dotenv",
    "setuptools",
    "wheel",
    "dashscope>=1.23.6",
    "gevent~=24.11.1",
    "tiktoken~=0.8.0",
    "openai~=1.96.1",
    "mem0ai~=0.1.114",
    "langchain-openai~=0.3.28",
    "aiomysql~=0.2.0",
    "asyncpg~=0.30.0",
    "pymysql~=1.1.0 ",
    "psycopg2-binary~=2.9.10",
    "sqlalchemy[asyncio]>=2.0.29",
    "uvicorn>=0.34.2",
    "redis",
    "langfuse~=3.9.0",
    "model_sdk @ file:///Users/james/daocloud/code/dac/expert-agent/sdks/model_sdk-0.1.0-py3-none-any.whl",

]

[project.scripts]
agent = "agent.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pyright]
include = ["agent"]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.uv]
dev-dependencies = []
index-url = "https://mirrors.aliyun.com/pypi/simple/"